+++ src/classes/babies/PeelingBaby.ts
@@ -1,0 +1,20 @@
+import { CollectibleType } from "isaac-typescript-definitions";
+import {
+  CallbackCustom,
+  ModCallbackCustom,
+  useActiveItemTemp,
+} from "isaacscript-common";
+import { Baby } from "../Baby";
+
+/** Pinking Shears effect on hit. */
+export class PeelingBaby extends Baby {
+  override isValid(player: EntityPlayer): boolean {
+    return !player.HasCollectible(CollectibleType.PINKING_SHEARS);
+  }
+
+  @CallbackCustom(ModCallbackCustom.ENTITY_TAKE_DMG_PLAYER)
+  entityTakeDmgPlayer(player: EntityPlayer): boolean | undefined {
+    useActiveItemTemp(player, CollectibleType.PINKING_SHEARS);
+    return undefined;
+  }
+}