+++ docs/history.md
@@ -6,6 +6,7 @@
 ### _v1.20.0_ - September 18th, 2023
 
 - The following babies have new effects:
+  - Peeling Baby (#199) - "Starts with Potato Peeler" --> "Pinking Shears effect on hit"
   - X Eyed Baby (#309) - "Starts with Marked" --> "Starts with 3x Lil Brimstone + Marked"
   - Dragon Baby (#432) - "Starts with Lil Brimstone" --> "Starts with Montezuma's Revenge + Playdough Cookie + My Reflection"
   - Penguin Baby (#554) - "Starts with Blood Puppy" --> "Starts with friendly level 3 Blood Puppy"